REFACTOR: Fix duplicated code in Set-Variable/Get-Variable commands
Introduce a VariableCommandBase class which is part of the official
PowerShell API. This class contains the wildcard filtering logic for
Include and Exclude parameters.
